requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
telethon>=1.36.0
pydantic>=2.7.0
python-dotenv>=1.0.0
